requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
psycopg2-binary>=2.7,<3.0
django-bootstrap4==0.0.7
django-filter==2.0.0
django-import-export==1.2.0
djangorestframework==3.9.1